Fix German translation of panel context menu.
The second header in panel's context menu has been translated as "Application menu" instead of "Panel".
Correcting here (GitHub) and not on pootle.lxde.org as it's one of those infamous country-specific ("_DE") files which don't exist on the latter at all (and which are utterly useless in German, IMO, which is a different story, though).